+++ src/main/java/org/plumelib/util/CollectionsPlume.java
@@ -110,7 +110,8 @@ public final class CollectionsPlume {
   }
 
   /**
-   * Returns a copy of the list with duplicates removed. Retains the original order.
+   * Returns a copy of the list (never the original list) with duplicates removed, but retaining the
+   * original order.
    *
    * @param <T> type of elements of the list
    * @param l a list to remove duplicates from
@@ -125,32 +126,62 @@ public final class CollectionsPlume {
   }
 
   /**
-   * Returns a copy of the list with duplicates removed. Retains the original order. May return its
-   * argument if its argument has no duplicates, but is not guaranteed to do so.
+   * Returns a copy of the list with duplicates removed, but retaining the original order. May
+   * return its argument if its argument has no duplicates, but is not guaranteed to do so. The
+   * argument is not modified.
    *
-   * <p>If the element type implements {@link Comparable}, use {@link #withoutDuplicatesComparable}.
+   * <p>If the element type implements {@link Comparable}, use {@link #withoutDuplicatesSorted} or
+   * {@link #withoutDuplicatesComparable}.
    *
    * @param <T> the type of elements in {@code values}
    * @param values a list of values
    * @return the values, with duplicates removed
    */
   public static <T> List<T> withoutDuplicates(List<T> values) {
-    HashSet<T> hs = new LinkedHashSet<>(values);
-    if (values.size() == hs.size()) {
+    Set<T> s = ArraySet.newArraySetOrLinkedHashSet(values);
+    if (values.size() == s.size()) {
       return values;
     } else {
-      return new ArrayList<>(hs);
+      return new ArrayList<>(s);
     }
   }
 
   /**
-   * Returns a list with the same contents as its argument, but without duplicates. May return its
-   * argument if its argument has no duplicates, but is not guaranteed to do so.
+   * Returns a list with the same contents as its argument, but sorted and without duplicates. May
+   * return its argument if its argument is sorted and has no duplicates, but is not guaranteed to
+   * do so. The argument is not modified.
    *
    * <p>This is like {@link #withoutDuplicates}, but requires the list elements to implement {@link
-   * Comparable}, and thus can be more efficient. Also, this does not retain the original order; the
-   * result is sorted.
+   * Comparable}, and thus can be more efficient.
    *
+   * @see #withoutDuplicatesComparable
+   * @param <T> the type of elements in {@code values}
+   * @param values a list of values
+   * @return the values, with duplicates removed
+   */
+  public static <T extends Comparable<T>> List<T> withoutDuplicatesSorted(List<T> values) {
+    // This adds O(n) time cost, and has the benefit of sometimes avoiding allocating a TreeSet.
+    if (isSortedNoDuplicates(values)) {
+      return values;
+    }
+
+    Set<T> set = new TreeSet<>(values);
+    return new ArrayList<>(set);
+  }
+
+  /**
+   * Returns a list with the same contents as its argument, but without duplicates. May return its
+   * argument if its argument has no duplicates, but is not guaranteed to do so. The argument is not
+   * modified.
+   *
+   * <p>This is like {@link #withoutDuplicatesSorted}, but it is not guaranteed to return a sorted
+   * list. Thus, it is occasionally more efficient.
+   *
+   * <p>This is like {@link #withoutDuplicates}, but requires the list elements to implement {@link
+   * Comparable}, and thus can be more efficient. If a new list is returned, this does not retain
+   * the original order; the result is sorted.
+   *
+   * @see #withoutDuplicatesSorted
    * @param <T> the type of elements in {@code values}
    * @param values a list of values
    * @return the values, with duplicates removed
@@ -212,7 +243,7 @@ public final class CollectionsPlume {
    * @return true if the list is sorted and has no duplicates
    */
   public static <T extends Comparable<T>> boolean isSortedNoDuplicates(List<T> values) {
-    if (values.isEmpty() || values.size() == 1) {
+    if (values.size() < 2) {
       return true;
     }